Add data packet watchdog
If no data packet is emitted for a configurable count of iterations, an
error will be raised, in turn crashing the engine (and potentially
triggering a restart).

The count of iterations can be configured by adding the
'max_no_data_iterations' option in the driver's section (default is 5).
An iteration will happen every 5 seconds when no data is received.